+++ src/user-service/src/use_cases/commands/mod.rs
@@ -7,3 +7,4 @@ pub mod delete_user;
 pub mod set_user_stripe_customer_id;
 pub mod update_access_token;
 pub mod update_user_profile;
+pub mod upsert_newsletter_subscription;